rename package as arctern (#7)
