+++ scala/src/main/scala/org/apache/spark/sql/arctern/UdtRegistratorWrapper.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.udt
+package org.apache.spark.sql.arctern
 
 import org.locationtech.jts.geom.Geometry
 import org.apache.spark.sql.types.UDTRegistration
